requirements.txt
rvbd_portal>=0.1

Django>=1.5.1,<1.6
django-model-utils==1.2.0
djangorestframework==2.3.8
django-extensions==1.1.1
flyscript>=0.5.6
jsonfield==0.9.5
numpy>=1.7.0
pandas>=0.10.1
pygeoip>=0.2.6
python-dateutil>=2.1
pytz>=2013.8
six>=1.3.0
wsgiref>=0.1.2
